Add kubelet-wrapper and etcd-wrapper
The kubelet-wrapper and etcd-wrapper scripts are part of Flatcar
Container Linux. They are tightly coupled with the service files
kubelet.service and etcd-member.service, respectively. The service files
are part of Lokomotive.

This coupling consists of environment variables and container runtime
parameters. This causes problem because any change (e.g. moving from rkt
to Docker) requires synchronisation between Flatcar and Lokomotive
releases.

This patch makes Lokomotive provide its own kubelet-wrapper and
etcd-wrapper scripts so that it is not dependent on the a specific
channel of Flatcar.

This patch also changes the container runtime to Docker.

Since Docker does not support sd_notify, I remove the related code.
To fix this, we could use sdnotify-proxy. See links:
https://www.freedesktop.org/software/systemd/man/sd_notify.html
https://github.com/coreos/sdnotify-proxy

In practice, Lokomotive works without sd_notify, but other services that
have a dependency on etcd-member might have not run in the correct
order. I found only one service in this case: flanneld:

```
/usr/lib/systemd/system/flanneld.service:4:After=etcd.service etcd2.service etcd-member.service
```

This patch only fixes AWS: packet and others are left as TODO.

I expect that kubelet-wrapper and etcd-wrapper can be shared for all
cloud providers, so I've put them in the 'common' directory. This will
make future changes easier to manage.

To make terraform modules self-sufficient (e.g.
aws/flatcar-linux/kubernetes), required files should normally not be
outside of the root directory of the terraform module (e.g. no access
via ../../../). I've added symbolic links to the common directory. If
you create assets with vfsgen, the symbolic links are read as regular
files so they get added several times in the assets, making the
terraform module self-sufficient.

Note: with this patch, we are dangerously close to the 16384 size limit
on AWS user_data. In my tests, we're around 15860.

TODO:
- Fix sd_notify for Flannel
- Other cloud providers
- Longer-term solution for the 16384 size limit on AWS user_data