[EXPERIMENTAL] packet: remove depedency on Route53.
Remove the needed for Route 53 and allow the user to setup the DNS entries by hand.

Terraform will print in the screen asking the user to setup 3 different DNS entries
that are needed for lokomotive, after those are ready the user should do

```
touch /tmp/dns1
touch /tmp/dns2
touch /tmp/dns3
```

to continue with the installation.